+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Improved the validation for `currency` in various endpoints
+- Harmonized the setting of a default locale in various components
 - Set the parser to `angular` in the `prettier` options
 
 ## 2.54.0 - 2024-02-19